requirements.txt
# General Modules
Django
djangorestframework

# Third-party Modules
markdown
django-filter
drf-yasg[validation]
Pillow

# Deployment Modules
django-cors-headers